requirements.txt
anyio==4.6.2.post1
APScheduler==3.9.1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
python-dotenv==1.0.0
python-telegram-bot==20.0
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
